This project was written in the course of the academic year 2018-2019 for the telecommunication physics course at ULB. It was asked to build a ray-tracing software calculating the received power in a building level to assess wi-fi coverage.
The proprietary Qt library was used to build intuitive GUI, and the software was initially build for ease of use
and extensive use of the entenna formalism.

Written by Mateusz Szydelko and myself, this project served as a basis for a Master thesis at the OPERA department
for 5G technology modelling, I was however, not involved past the foundational work.

Not to be used, for demonstration purpose only.